requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.3.0
python-dotenv>=0.19.0
yfinance>=0.2.0 